@@ -12,12 +12,25 @@
         }
     ],
     "require": {
-        "php": ">=7.4.0",
-        "laravel/framework": ">=6.0"
+        "php": ">=8.0.0",
+        "laravel/framework": ">=8.0"
+    },
+    "require-dev": {
+        "pestphp/pest": "^2.0",
+        "phpunit/phpunit": "^9.5|^10.0",
+        "mockery/mockery": "^1.5"
     },
     "autoload": {
         "psr-4": {
             "LeMaX10\\SimpleActions\\": "src/"
+        },
+        "files": [
+            "src/helpers.php"
+        ]
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "LeMaX10\\SimpleActions\\Tests\\": "tests/"
         }
     },
     "extra": {
@@ -25,6 +38,14 @@
             "providers": [
                 "LeMaX10\\SimpleActions\\SimpleActionServiceProvider"
             ]
+        }
+    },
+    "scripts": {
+        "test": "vendor/bin/pest"
+    },
+    "config": {
+        "allow-plugins": {
+            "pestphp/pest-plugin": true
         }
     }
 }
