@@ -6,14 +6,49 @@ namespace LeMaX10\SimpleActions\Traits;
 use Illuminate\Contracts\Events\Dispatcher;
 use Illuminate\Events\NullDispatcher;
 use Illuminate\Events\QueuedClosure;
+use LeMaX10\SimpleActions\Events\ActionAfterRun;
+use LeMaX10\SimpleActions\Events\ActionBeforeRun;
+use LeMaX10\SimpleActions\Events\ActionFailed;
+use LeMaX10\SimpleActions\Events\ActionRan;
+use LeMaX10\SimpleActions\Events\ActionRunning;
+use LeMaX10\SimpleActions\Observers\ActionObserver;
 
 trait HasEvents
 {
-    protected $dispatchesEvents = [];
+    protected array $dispatchesEvents = [];
 
-    protected static Dispatcher|null $dispatcher = null;
+    protected static ?Dispatcher $dispatcher = null;
+
+    protected static array $observers = [];
 
     protected const EVENT_PREFIX = 'simple-action';
+
+    protected static array $eventMap = [
+        'beforeRun' => ActionBeforeRun::class,
+        'running' => ActionRunning::class,
+        'ran' => ActionRan::class,
+        'failed' => ActionFailed::class,
+        'afterRun' => ActionAfterRun::class,
+    ];
+
+    /**
+     * @param  ActionObserver|string  $observer
+     * @return void
+     */
+    public static function observe(ActionObserver|string $observer): void
+    {
+        $observerInstance = is_string($observer) ? app($observer) : $observer;
+
+        static::$observers[static::class][] = $observerInstance;
+
+        foreach (static::$eventMap as $event => $eventClass) {
+            if (method_exists($observerInstance, $event)) {
+                static::registerActionEvent($event, function ($eventObject) use ($observerInstance, $event) {
+                    return $observerInstance->{$event}($eventObject);
+                });
+            }
+        }
+    }
 
     /**
      * @param  string  $event
@@ -29,10 +64,11 @@ trait HasEvents
 
     /**
      * @param  string  $event
+     * @param  array  $args
      * @param  bool  $halt
      * @return mixed
      */
-    protected function fireActionEvent(string $event, bool $halt = true): mixed
+    protected function fireActionEvent(string $event, array $args = [], bool $halt = true): mixed
     {
         if (empty(static::$dispatcher)) {
             return true;
@@ -40,31 +76,63 @@ trait HasEvents
 
         $method = $halt ? 'until' : 'dispatch';
 
+        $eventObject = $this->createEventObject($event, $args);
+
         $result = $this->filterActionEventResults(
-            $this->fireCustomActionEvent($event, $method)
+            $this->fireCustomActionEvent($event, $method, $eventObject)
         );
 
         if ($result === false) {
             return false;
         }
 
-        return ! empty($result) ? $result : static::$dispatcher->{$method}(
-            static::prefixable($event .': '. static::class), $this
+        $globalResult = static::$dispatcher->{$method}(
+            static::prefixable($event .': '. static::class),
+            $eventObject
         );
+
+        if ($globalResult === false) {
+            return false;
+        }
+
+        return ! empty($result) ? $result : $globalResult;
+    }
+
+    /**
+     * @param  string  $event
+     * @param  array  $args
+     * @return object
+     */
+    protected function createEventObject(string $event, array $args = []): object
+    {
+        if (isset($this->dispatchesEvents[$event])) {
+            return new $this->dispatchesEvents[$event]($this, ...$args);
+        }
+
+        if (isset(static::$eventMap[$event])) {
+            $eventClass = static::$eventMap[$event];
+            return new $eventClass($this, ...$args);
+        }
+
+        //fallbacck
+        return new class($this, $args) extends ActionEvent {
+            public function __construct(public $action, public array $arguments = []) {}
+        };
     }
 
     /**
      * @param  string  $event
      * @param  string  $method
+     * @param  object  $eventObject
      * @return mixed|void
      */
-    protected function fireCustomActionEvent(string $event, string $method)
+    protected function fireCustomActionEvent(string $event, string $method, object $eventObject)
     {
         if (! isset($this->dispatchesEvents[$event])) {
             return;
         }
 
-        $result = static::getEventDispatcher()->$method(new $this->dispatchesEvents[$event]($this));
+        $result = static::getEventDispatcher()->$method($eventObject);
 
         if (! is_null($result)) {
             return $result;
@@ -99,6 +167,33 @@ trait HasEvents
      * @param  QueuedClosure|callable|array|class-string  $callback
      * @return void
      */
+    public static function running(QueuedClosure|callable|array|string $callback): void
+    {
+        static::registerActionEvent('running', $callback);
+    }
+
+    /**
+     * @param  QueuedClosure|callable|array|class-string  $callback
+     * @return void
+     */
+    public static function ran(QueuedClosure|callable|array|string $callback): void
+    {
+        static::registerActionEvent('ran', $callback);
+    }
+
+    /**
+     * @param  QueuedClosure|callable|array|class-string  $callback
+     * @return void
+     */
+    public static function failed(QueuedClosure|callable|array|string $callback): void
+    {
+        static::registerActionEvent('failed', $callback);
+    }
+
+    /**
+     * @param  QueuedClosure|callable|array|class-string  $callback
+     * @return void
+     */
     public static function afterRun(QueuedClosure|callable|array|string $callback): void
     {
         static::registerActionEvent('afterRun', $callback);
@@ -113,9 +208,13 @@ trait HasEvents
             return;
         }
 
-        $instance = new static;
-        foreach ($instance->dispatchesEvents as $event) {
-            static::getEventDispatcher()->forget($event);
+        foreach (static::$eventMap as $event => $eventClass) {
+            static::getEventDispatcher()->forget(static::prefixable($event .': '. static::class));
+        }
+
+        // Очищаем observers
+        if (isset(static::$observers[static::class])) {
+            unset(static::$observers[static::class]);
         }
     }
 
@@ -130,7 +229,7 @@ trait HasEvents
     /**
      * @return Dispatcher|null
      */
-    public static function getEventDispatcher(): Dispatcher|null
+    public static function getEventDispatcher(): ?Dispatcher
     {
         return static::$dispatcher;
     }
@@ -139,7 +238,7 @@ trait HasEvents
      * @param  Dispatcher|null  $dispatcher
      * @return void
      */
-    public static function setEventDispatcher(Dispatcher|null $dispatcher): void
+    public static function setEventDispatcher(?Dispatcher $dispatcher): void
     {
         static::$dispatcher = $dispatcher;
     }
