@@ -3,7 +3,12 @@ declare(strict_types=1);
 
 namespace LeMaX10\SimpleActions\Exceptions;
 
-class ActionHandlerMethodNotFoundException extends \ApplicationException
+/**
+ * Исключение, которое выбрасывается, когда метод handle не найден в действии
+ * 
+ * @author Vladimir Pyankov, v@pyankov.pro, RDLTeam 
+ */
+class ActionHandlerMethodNotFoundException extends \Exception
 {
     public function __construct(string $class)
     {
